extracted useEffect into fetchPromptPageData
